Improvements for HTTP CGI server
This submission makes the following changes to the PASS HTTP CGI server:
* The response header allows all callers to receive responses.  This
relates to the pass-ember submission for issue 54.
* Exceptions resulting from the execution of the server's command are
now caught and returned as cod 500.
* The arguments specifcation mechanism how allows an array of arguments
to be specified, passed to the command and printed in case of error.
The default command is now the one given to me by Elliot on 12/11/17.